requirements.txt
beautifulsoup4==4.9.3
lxml==4.6.3
numpy==1.21.4
opencv-python==4.5.4.60
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.4.0
pytz==2021.3
requests==2.25.1
selenium==4.1.0
Selenium-Screenshot==1.7.0
shutils==0.1.0
urllib3==1.26.7
user-agent==0.1.9
yarl==1.7.2
wheel==0.37.0
